Build.Drasil

Description

Re-export all contents in drasil-build for external use.

Makefiles

Classes

class RuleTransformer c where

Members can be transformed into a make rule.

Methods

makeRule :: c -> [Rule]

Types and Functions

type Annotation = [Comment]

A Makefile Annotation is made of 0 or more Comments

data Command

A command is made up of MakeStrings and command operators.

mkCheckedCommand :: MakeString -> Command

Creates a Command which fails the make process if it does not return zero.

mkCommand :: MakeString -> Command

Creates a command which executes and ignores the return code.

type Dependencies = [Target]

Dependencies are made up of 0 or more Targets.

data Rule

A Makefile Rule can have comments and commands but needs a target, dependencies, and a type.

mkFile :: Annotation -> Target -> Dependencies -> [Command] -> Rule

Creates a Rule which results in a file being created.

mkRule :: Annotation -> Target -> Dependencies -> [Command] -> Rule

Creates an abstract Rule not associated to a specific file.

data MakeString

(+:+) :: MakeString -> MakeString -> MakeString

Concatenates two MakeStrings with a space in between.

makeS :: String -> MakeString

Constructor for converting a String into a MakeString.

mkFreeVar :: VarName -> MakeString

Constructor for Free variables.

mkImplicitVar :: VarName -> MakeString

Constructor for Implicit variables.

mkWindowsVar :: VarName -> VarVal -> VarVal -> MakeString

Constructor for Windows OS variables.

mkOSVar :: VarName -> VarVal -> VarVal -> VarVal -> MakeString

Constructor for OS variables.

Printers

genMake :: RuleTransformer c => [c] -> Doc

Generates the makefile by calling build after toMake.