Build.Drasil.Make.AST

Description

Defines a Makefile abstract syntax tree.

Types

newtype Makefile

A Makefile is made up of Makefile rules.

Constructors

M [Rule]

data Rule

A Makefile Rule can have comments and commands but needs a target, dependencies, and a type.

Constructors

R Annotation Target Dependencies Type [Command]

data Command

A command is made up of MakeStrings and command operators.

Constructors

C MakeString [CommandOpts]

data CommandOpts

Ignore the return code from the system.

Constructors

IgnoreReturnCode

data Type

Type of rule, either abstract or file-oriented.

Constructors

Abstract

File

type Annotation = [Comment]

A Makefile Annotation is made of 0 or more Comments

type Target = MakeString

A Makefile target is made from a MakeString.

type Dependencies = [Target]

Dependencies are made up of 0 or more Targets.

Constructors

mkFile :: Annotation -> Target -> Dependencies -> [Command] -> Rule

Creates a Rule which results in a file being created.

mkRule :: Annotation -> Target -> Dependencies -> [Command] -> Rule

Creates an abstract Rule not associated to a specific file.

mkCheckedCommand :: MakeString -> Command

Creates a Command which fails the make process if it does not return zero.

mkCommand :: MakeString -> Command

Creates a command which executes and ignores the return code.